Refactor and enhance runtime components
- Added missing include for prefix.hpp in msvcrt.cpp.
- Updated copy operations in psapi.cpp to use static_cast for better type safety.
- Cast result of gui::message_box to int in user32.cpp to ensure proper return type.
- Included vector header in ws2_32.cpp for potential future use.
- Introduced legacy_winapi.cpp to implement various Windows API functions, including file handling, process creation, and memory management, with proper error handling and path normalization.

+++ src/runtime/comdlg32.cpp
@@ -76,7 +76,7 @@ TL_COMDLG_MSABI int tl_GetOpenFileNameW(void* open_filename) noexcept {
         if (ofn->file[0] == 0) {
             const std::u16string u16 = util::utf8_to_wide("C:\\document.txt");
             const std::size_t len = std::min<std::size_t>(u16.size(), ofn->max_file - 1);
-            std::copy(u16.begin(), u16.begin() + len, ofn->file);
+            std::copy(u16.begin(), u16.begin() + static_cast<std::ptrdiff_t>(len), ofn->file);
             ofn->file[len] = 0;
         }
         return 1;
